Validations Added.
I added some validations to the code. Probably there are many validations missing but when I detect them, they will be added soon.

+++ backend/todo/adapter/database/mongo.py
@@ -39,16 +39,24 @@ class Mongo(Database):
         """Return list of Todos that comply with the received filter."""
         if filter_by.lower() == "completed":
             todos = list(self.collection.find({"task.completed": True}))
+        elif filter_by.lower() == "uncompleted":
+            todos = list(self.collection.find({"task.completed": False}))
+
+        if todos:
             return todos
 
-        if filter_by.lower() == "uncompleted":
-            todos = list(self.collection.find({"task.completed": False}))
-            return todos
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND, detail="Todos not found."
+        )
 
     def all(self) -> list[Todo]:
         """Get and return a list of Todos in the system."""
         todos = list(self.collection.find())
-        return todos
+
+        if todos:
+            return todos
+
+        raise HTTPException(status_code=status.HTTP_204_NO_CONTENT)
 
     def get(self, id: str) -> Todo:
         """Get and return a specific Todo with the given ID."""
@@ -64,13 +72,17 @@ class Mongo(Database):
     def update(self, id: str, todo: Todo) -> Todo:
         """Update and return a Todo with the given ID using the new data."""
         new_data = todo.dict(exclude_unset=True)
-        self.collection.find_one_and_update(
-            {"id": id},
-            {"$set": new_data},
-            return_document=ReturnDocument.AFTER,
-        )
-        return todo
+
+        if self.get(id):
+            self.collection.find_one_and_update(
+                {"id": id},
+                {"$set": new_data},
+                return_document=ReturnDocument.AFTER,
+            )
+            return todo
 
     def delete(self, id: str) -> None:
         """Delete a single Todo with the given ID."""
-        self.collection.delete_one({"id": id})
+        if self.get(id):
+            self.collection.delete_one({"id": id})
+            HTTPException(status_code=status.HTTP_204_NO_CONTENT)
